refactor: รวมฟังก์ชัน box ops
