+++ src/packages/models/yolov8/yolov8onnx/yolov8object/utils.py
@@ -1,6 +1,8 @@
 import numpy as np
 import cv2
 
+from models.common.box_ops import compute_iou, nms, multiclass_nms, xywh2xyxy
+
 class_names = ['person', 'bicycle', 'car', 'motorcycle', 'airplane', 'bus', 'train', 'truck', 'boat', 'traffic light',
                'fire hydrant', 'stop sign', 'parking meter', 'bench', 'bird', 'cat', 'dog', 'horse', 'sheep', 'cow',
                'elephant', 'bear', 'zebra', 'giraffe', 'backpack', 'umbrella', 'handbag', 'tie', 'suitcase', 'frisbee',
@@ -16,71 +18,6 @@ np.random.seed(3)
 colors = np.random.uniform(0, 255, size=(100, 3))
 
 
-def nms(boxes, scores, iou_threshold):
-    # Sort by score
-    sorted_indices = np.argsort(scores)[::-1]
-
-    keep_boxes = []
-    while sorted_indices.size > 0:
-        # Pick the last box
-        box_id = sorted_indices[0]
-        keep_boxes.append(box_id)
-
-        # Compute IoU of the picked box with the rest
-        ious = compute_iou(boxes[box_id, :], boxes[sorted_indices[1:], :])
-
-        # Remove boxes with IoU over the threshold
-        keep_indices = np.where(ious < iou_threshold)[0]
-
-        # print(keep_indices.shape, sorted_indices.shape)
-        sorted_indices = sorted_indices[keep_indices + 1]
-
-    return keep_boxes
-
-def multiclass_nms(boxes, scores, class_ids, iou_threshold):
-
-    unique_class_ids = np.unique(class_ids)
-
-    keep_boxes = []
-    for class_id in unique_class_ids:
-        class_indices = np.where(class_ids == class_id)[0]
-        class_boxes = boxes[class_indices,:]
-        class_scores = scores[class_indices]
-
-        class_keep_boxes = nms(class_boxes, class_scores, iou_threshold)
-        keep_boxes.extend(class_indices[class_keep_boxes])
-
-    return keep_boxes
-
-def compute_iou(box, boxes):
-    # Compute xmin, ymin, xmax, ymax for both boxes
-    xmin = np.maximum(box[0], boxes[:, 0])
-    ymin = np.maximum(box[1], boxes[:, 1])
-    xmax = np.minimum(box[2], boxes[:, 2])
-    ymax = np.minimum(box[3], boxes[:, 3])
-
-    # Compute intersection area
-    intersection_area = np.maximum(0, xmax - xmin) * np.maximum(0, ymax - ymin)
-
-    # Compute union area
-    box_area = (box[2] - box[0]) * (box[3] - box[1])
-    boxes_area = (boxes[:, 2] - boxes[:, 0]) * (boxes[:, 3] - boxes[:, 1])
-    union_area = box_area + boxes_area - intersection_area
-
-    # Compute IoU
-    iou = intersection_area / union_area
-
-    return iou
-
-
-def xywh2xyxy(x):
-    # Convert bounding box (x, y, w, h) to bounding box (x1, y1, x2, y2)
-    y = np.copy(x)
-    y[..., 0] = x[..., 0] - x[..., 2] / 2
-    y[..., 1] = x[..., 1] - x[..., 3] / 2
-    y[..., 2] = x[..., 0] + x[..., 2] / 2
-    y[..., 3] = x[..., 1] + x[..., 3] / 2
-    return y
 
 
 # def draw_detections(image, boxes, scores, class_ids, mask_alpha=0.3):
